Rename scenario to vm scheduling
