+++ examples/data_center/best_fit/config.yml
@@ -1,5 +1,5 @@
 env:
-  scenario: data_center
+  scenario: vm_scheduling
   topology: azure.2019.10k
   start_tick: 0
   durations: 8638
