+++ examples/data_center/random/config.yml
@@ -1,5 +1,5 @@
 env:
-  scenario: data_center
+  scenario: vm_scheduling
   topology: azure.2019.original
   start_tick: 0
   durations: 8638
