+++ maro/simulator/scenarios/vm_scheduling/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .business_engine import DataCenterBusinessEngine
+from .business_engine import VmSchedulingBusinessEngine
 from .common import (
     Action, AssignAction, DecisionPayload, Latency, PostponeAction, PostponeType, ValidPhysicalMachine,
     VmFinishedPayload, VmRequestPayload
@@ -12,7 +12,7 @@ from .physical_machine import PhysicalMachine
 from .virtual_machine import VirtualMachine
 
 __all__ = [
-    "DataCenterBusinessEngine",
+    "VmSchedulingBusinessEngine",
     "Action", "AssignAction", "PostponeAction",
     "DecisionPayload",
     "Latency",
