+++ maro/simulator/scenarios/vm_scheduling/meta/cpu_readings.yml
